Updating ESLint rules; Adding React Router to project; Refactored App and created Home component; Refactored Login and Logout to handle login and logout with React Router; Created routerContext to help with routing
