+++ package.json
@@ -23,7 +23,8 @@
     "@inrupt/vocab-common-rdf": "^1.0.5",
     "fs-extra": "^11.1.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.10.0"
   },
   "browserslist": [
     "last 3 and_chr versions",
