+++ src/components/AppHeader.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const AppHeader = () => (
+  <header>
+    <h2>Getting Started with PASS</h2>
+  </header>
+);
+
+export default AppHeader;
