+++ src/components/Login/Login.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { LoginButton } from '@inrupt/solid-ui-react';
 import { SOLID_IDENTITY_PROVIDER } from '../../utils';
 
@@ -10,32 +10,24 @@ import { SOLID_IDENTITY_PROVIDER } from '../../utils';
  * @name Login
  */
 
-const Login = () => {
-  const [currentUrl, setCurrentUrl] = useState('http://localhost:3000');
-
-  useEffect(() => {
-    setCurrentUrl(window.location.href);
-  }, [setCurrentUrl]);
-
-  return (
-    <section id="login" className="panel">
-      <div className="row">
-        <label id="labelLogin" htmlFor="btnLogin">
-          Click the following login button to log into your pod at [
-          <a href={SOLID_IDENTITY_PROVIDER} target="_blank" rel="noreferrer">
-            {SOLID_IDENTITY_PROVIDER}
-          </a>
-          ]:{' '}
-        </label>
-        <LoginButton
-          oidcIssuer={SOLID_IDENTITY_PROVIDER}
-          redirectUrl={currentUrl}
-          onError={console.error}
-        />
-        <p>Not logged in</p>
-      </div>
-    </section>
-  );
-};
+const Login = ({ currentUrl }) => (
+  <section id="login" className="panel">
+    <div className="row">
+      <label id="labelLogin" htmlFor="btnLogin">
+        Click the following login button to log into your pod at [
+        <a href={SOLID_IDENTITY_PROVIDER} target="_blank" rel="noreferrer">
+          {SOLID_IDENTITY_PROVIDER}
+        </a>
+        ]:{' '}
+      </label>
+      <LoginButton
+        oidcIssuer={SOLID_IDENTITY_PROVIDER}
+        redirectUrl={`${currentUrl}home/`}
+        onError={console.error}
+      />
+      <p>Not logged in</p>
+    </div>
+  </section>
+);
 
 export default Login;
