+++ src/contexts/routerContext.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const RouterContext = createContext('');
+
+export default RouterContext;
